RFNoC: Cache and re-use host endpoints
Constatntly incrementing endpoints was causing the entries in the
routing table on the device to be exhausted, eventually resulting in a
timeout error on control packets.  Since a connection between the host
and a stream endpoint on a device in a given direction is unique, the
host endpoints can be cached and re-used.  This change does that.

Signed-off-by: michael-west <michael.west@ettus.com>