uhd: Demote WARNING on minor compat mismatch to DEBUG
When the minor FPGA compat number on the device is ahead of what MPM
expects, we no longer print a warning. That's because by definition, the
FPGA is still compatible with the software in this case.

If UHD requires a certain minor compat number to enable a feature,
the appropriate behaviour would be to print a warning only for that
case.

This is the equivalent change to the MPM-only change in 88d28481.