tools: uhd_ubuntu_deb: Require --buildpath
Because this script deletes the entire contents of --buildpath, we no
longer provide '..' as the default. First, it is not a suitable default,
because it will contain the directory from which this script is called.
Second, if someone runs this script from a source repo for UHD, and this
repo is checked out into a non-empty directory, the script will still,
by default, delete all the contents from that directory.

Now, the script fails when not providing --buildpath. Note that CI code
is already using this command line argument.