n310: cpld: Get and set TX ATR bits
This allows the user to get the current state of the tx atr bits and set
them back to a given state. This is useful for the n310 when resetting
the front end, in order to avoid any tx power out of the frontend when
the init_cals are run.

Signed-off-by: mattprost <matt.prost@ni.com>