multi_usrp: Shorten GPIO bank names on single-db USRPs
In particular, the E3xx series have a single "daughterboard", and
a single radio block with two channels. Prior to this fix,
multi_usrp::get_gpio_banks() would return "INT0A", where the 'A' is to
distinguish the INT0 bank with other dboards. Since there are none, the
'A' is unnecessary.