debian: Resolve lintian issues
This PR makes a couple of minor changes to Debian packaging files to
resolve errors reported by Lintian during CI and manual building of
.debs after having been updated in 38fb5e262:

- Remove instances of `${python3:Depends}` in the rules file, as this
substitution directive cannot be used without specifying `--with
python3` in the `dh` debhelper command invocation. However, the Docker
containers that are spun up to build .debs as part of UHD CI do not
include the Debian Python helper package necessary to use the
substitution even if `--with python3` is specified. Thus, the
substitution was removed and replaced with the actual names of the
dependent Python packages.

- Remove the `Files-Excluded` section from the copyright file. Lintian
flags an error whenever a file in that section is present in the source
tree (which the specified files are). The addition of this section was
supposed to bring UHD's built-in Debian packaging files more in line
with those used by the actual Debian package maintainer, but it is
likely that the actual Debian packing process runs additional steps that
reference this section and uses it to remove the specified files which
we currently don't run.