cmake: Fix linking DPDK when installed at non-standard location
With DPDK 21.11 installed at a non-standard path, the UHD build process failed
to find its shared libraries (librte_*.so) while linking libuhd.so. The
-l flags were passed to the linker as bare filenames, but the -L flag was not
passed to specify the library search path (although that information was
provided to cmake by pkg-config). This patch uses the full path to DPDK
libraries at link time instead of bare filenames, which fixes this case.