rfnoc: Add topology graph object
This is a class that stores the topology of RFNoC devices in a session
in a graph object. Nodes in such a graph include crossbars, transport
adapters, and stream endpoints. They are statically defined by the
underlying hardware.

This graph is different from the detail::graph, which stores the
configuration of the RFNoC blocks that the user chooses at runtime.

The main purpose of this graph is to keep track of and manage available
routes within the RFNoC devices. The intention of this object is that
every rfnoc_graph owns exactly one of these graphs, where it can store
all the routing information.