rfnoc: Rename `skip_property_propagation` to `is_back_edge`
`skip_property_propagation` was always a bad choice for a name, but
since 3930a615, it is completely wrong.

In the C++ API, there is no change in API or ABI, because the argument
name is not part of API/ABI, although the documentation strings were
modified.

In the Python API, the argument name and default value are also modified
to match the C++ API. Scripts using named arguments may have to be
modified to track this change.