rfnoc: Transition stream managers and mgmt_portal to topo_graph_t
This is an invasive change in how the management portal stores routes.
Previously, it would store all the shortest routes from its respective
local SEP to all the nodes in the graph. The key change here is to
replace mgmt_portal_impl::_node_addr_map with mgmt_portal_impl::_tgraph.

However, this has a lot of downstream effects. mgmt_portal_impl itself
required a lot of changes to make use of _tgraph instead of
_node_addr_map. Some changes to graph_stream_manager.cpp and
link_stream_manager.cpp were also required. Note that there is a single
instance of a topology graph within an rfnoc_graph, and it is owned by
the graph_stream_manager.
A less invasive change was the removal of the _epid_addr_map cache
object, since the topology graph stores EPIDs itself, it was a redundant
data structure.

This change by itself offers no new features or functionality, and has
no API changes beyond the graph stream manager. However, it is the basis
for enabling more features down the line.