experts: Move expert framework into public API
To enable third party components (e.g. frontend modules) to use the
experts framework the headers need to be placed in the public API.
This change does this and fixes all dependencies. The experts
framework has been kept as internal API to enable changes more
easily, however it has been there without major changes for quite
a while already and therefore it should be safe to publish it.