tune_map: Publish and simplify zbx_tune_map
The tune map for ZBX included some mixer values which were redundant and
which didn't have a self-explaining name. This commit improves the
naming and makes the zbx_tune_map_item_t a public type so it can be
accessed from non-UHD applications because it is in the property tree
already anyway. While removing the redundancy some calculations in the
ZBX expert have been rewritten to be more readable.