rfnoc: Move detection of TAs post-LSM-init
This changes the order of operations for detecting transport adapters
that are not connected to UHD. The new order of operations is:
- Initialize every link stream manager (LSM) with a full topology
  discovery (this discovery is unchanged in this commit)
- After *all* link stream managers are initialized, query USRPs for
  transport adapters (TAs) that were not detected by discovery. This
  step is critical for enabling remote raw UDP streaming from/to
  adapters that are not connected to UHD (e.g., running UHD on a USRP
  and streaming raw UDP data out of an SFP port).

This avoids early out-of-discovery addition of TAs to the topology
graph. When using addr,second_addr for instance, this would cause the
first link to be established, and then the transport adapter for the
second link to be added to the graph prematurely, before second_addr
would be established (which also would have detected the second TA). The
problem is, this would skip an initialization step of the second TA.