MPM: Better error message for multiple assigned MACs
Each MAC address should be only assigned to one link at a time.
If the motherboard EEPROM is messed up, it might happen that
there are more links connected to a single MAC.

So in that case a

  [link_index] = ip2.link_lookup(address=mac_addr)

results in a ValueError (too many values to unpack) which gives
no clue on what actually went wrong. Instead we check the len
of he link_lookup result and give a more meaningful error message.