uhd: python: Make DeviceAddr behave like dict
This adds methods to the Python-bound version of device_addr_t so it
behaves like and can be used like a Python dictionary.

Specifically, it adds (including unit tests; assume a and b are of type
uhd.types.DeviceAddr):

```py
len(a)
a.keys()
a.values()
a.get('key')
a.get('key', 'default')
a.pop('key')
'key' in a
a['key']
a['key'] = 'new_val'
a == b
a != b
    # These differ from Python's versions:
a.update(b)
a.update({'foo': 'bar'})
```

The update() method follows the signature of
uhd::device_addr_t::update(), not that of Python's dict.update(), see
the manual for details (the fail_on_conflict argument). All other
methods behave equivalently between Python's dict and uhd::dict.

Signed-off-by: Martin Braun <martin.braun@ettus.com>