mpm: x4xx: Move get_master_clock_rate() to DB-RPC
The X4x0 device is an outlier with respect to all other MPM devices when
it comes to this API. All other MPM-devices define this API on the
daughterboard, not the motherboard.

For all the reasons laid out in dboard_manager/base.py, we move this API
call for X4x0 also into the daughterboard RPC space.

Both get_master_clock_rate() and set_master_clock_rate() are explicitly
declared @no_rpc in MPM.

This is an API breakage in the UHD/MPM communication API.